requirements.txt
alembic
uvicorn
strawberry-graphql
starlette
SQLAlchemy
pandas